test(webhook): separate auto-plan tests from explicit plan-command tests
The plan integration file mixed two distinct flows: the explicit /schemabot
plan command and automatic plan-on-push (PR opened/reopened/synchronize).
Move the auto-plan tests into auto_plan_integration_test.go, leaving the
plan-command tests in plan_integration_test.go. Pure code movement, no test
logic changes.
